完成历史记录功能

+++ app/src/main/java/com/example/administrator/essim/utils/LocalData.java
@@ -13,7 +13,7 @@ import java.util.List;
 
 public class LocalData {
 
-    private static final int HISTORY_SIZE = 5;
+    private static final int HISTORY_SIZE = 100;
     /**
      * 登陆成功后将账户信息保存到本地
      *
@@ -76,15 +76,12 @@ public class LocalData {
         List<ViewHistory> isSaved = DataSupport.where("illust_id=?",
                 String.valueOf(illustsBean.getId())).find(ViewHistory.class);
         if (isSaved != null && isSaved.size() != 0) {
-            /*isSaved.get(0).setView_time(Common.getTime(String.valueOf(System.currentTimeMillis())));
-            isSaved.get(0).save();*/
             ViewHistory book = new ViewHistory();
             book.setView_time(System.currentTimeMillis());
             book.updateAll("illust_id=?", String.valueOf(illustsBean.getId()));
         } else {
-            List<ViewHistory> list = new ArrayList<>();
-            list.addAll(DataSupport.order("view_time desc").find(ViewHistory.class));
-            if(list.size() >= HISTORY_SIZE){
+            List<ViewHistory> list = DataSupport.order("view_time desc").find(ViewHistory.class);
+            if(list != null && list.size() >= HISTORY_SIZE){
                 list.get(list.size() - 1).delete();
             }
             ViewHistory viewHistory = new ViewHistory(
